+++ client/modules/IDE/actions/toast.js
@@ -6,8 +6,24 @@ export function hideToast() {
   };
 }
 
-export function showToast(time) {
+/**
+ * Temporary fix until #2206 is merged.
+ * Supports legacy two-action syntax:
+ *    dispatch(setToastText('Toast.SketchFailedSave'));
+ *    dispatch(showToast(1500));
+ * And also supports proposed single-action syntax with message and optional timeout.
+ *    dispatch(showToast('Toast.SketchFailedSave'));
+ *    dispatch(showToast('Toast.SketchSaved', 5500));
+ */
+export function showToast(textOrTime, timeout = 1500) {
   return (dispatch) => {
+    let time = timeout;
+    if (typeof textOrTime === 'string') {
+      // eslint-disable-next-line no-use-before-define
+      dispatch(setToastText(textOrTime));
+    } else {
+      time = textOrTime;
+    }
     dispatch({
       type: ActionTypes.SHOW_TOAST
     });
